+++ .firebase/hosting.YnVpbGQvd2Vi.cache
@@ -1,5 +1,4 @@
 .DS_Store,1749936754528,5ee2931705ab4e076775fd99c630f01728ef7531aadba34619bf108cd6ea4405
-.last_build_id,1755147476086,f2403be18940145a0c531d9bb83166b4f10a0d6b482f9713407cd3af781777fe
 favicon.png,1749937016727,356c422dbee7c80fcd0de7f1f31bfd630dcdf1c135594b0b67e4ed97dc7f9fdf
 flutter.js,1752605912000,053adb95c9fd7e93f05fde87a56d68d3269388c2678826834ffd3ed7d292bf26
 launcher.png,1749936958127,f1b651238a58fe290baec6c5e32f3bdb1943dd2bd582f02569231f7a757c7837
@@ -21,8 +20,6 @@ assets/assets/sell02.jpeg,1749934233853,937bb7c3240b1725eb2bd79a759282cba71caf96
 assets/assets/sell05.jpeg,1749934233854,93cbe877da12298104f902c6d6a84e6564b6382fdb251d4fedebc5adf4842cdc
 assets/assets/sell06.jpeg,1749934233855,233a8e1d3806659c517724c831aa900ff87ed14456fcf0341d7fd5c135967386
 assets/assets/stock.jpeg,1749934233855,a641fd4a62b5fd371471b8a6786dec40872564b4af54854eba9b5ff40a0acafa
-assets/packages/simple_barcode_scanner/assets/barcode.html,1755145614112,47794e42afe19baf07f6f2511b0a92d0e0c0a644edf8e59a330a11fff145d557
-assets/packages/simple_barcode_scanner/assets/html5-qrcode.min.js,1755145614110,2677fd733829edeb6b0bdc9380da3dfac25df60239cd446cc98f2403f69b9efa
 canvaskit/canvaskit.js,1752606092000,9b87a0d6c67ac68cbce3c8018ecc18f7d77811460b772cd0c51bba45d120136c
 canvaskit/canvaskit.js.symbols,1752606090000,c58c818ceea58799cc5cfd951fa78f7607db6756342bd7a1aaeb1e431063894a
 canvaskit/canvaskit.wasm,1752606090000,c6fe626af107c4e97739d6c3062cfcb1f8e2eb75d8f1f657fdf77ac78b4dd934
@@ -36,16 +33,17 @@ icons/Icon-192.png,1749937016755,6d852e5bc012b05ea86166cdfa69b6cebc0eb70917edcc9
 icons/Icon-512.png,1749937016785,b86e81ab7345f118122b794ba09d4fdb9938cbd62591adea17f5cb1ba35034a0
 icons/Icon-maskable-192.png,1749937016799,6d852e5bc012b05ea86166cdfa69b6cebc0eb70917edcc95ced24b28a57244df
 icons/Icon-maskable-512.png,1749937016827,b86e81ab7345f118122b794ba09d4fdb9938cbd62591adea17f5cb1ba35034a0
-index.html,1755149608399,78df7e9ff716504616fa85a15da1f89bf925be5fc6fbca20eaf4b87827b2c872
-flutter_service_worker.js,1755149638543,4cdb1bca7a55f3cf6bf347f197a15104bbac19269962f831d024315d3de1cb79
-flutter_bootstrap.js,1755149608392,6577308640e460901191404be9a040523d972a8e27e629c767631e557a2d2a36
-version.json,1755149636311,263c49aaa0e5829f5b7107439b92217c03df18bb3b222c9bee89db582fc5ac00
-assets/AssetManifest.bin,1755149636471,19efd8e5b17731c1ebf6eeb2cdaabe09a0add7c9bb8a03e34ae0886880af992f
-assets/AssetManifest.bin.json,1755149636471,48623b11b920af55833fe0fb9c5c6324218534ff96c880be82e1f63996710baf
-assets/AssetManifest.json,1755149636471,cf608b15af663e17b1858b5c7bdfe5ea17fc95d73bf503cf54559b9834b62eac
-assets/FontManifest.json,1755149636471,9ea504185602e57d97b7c3517d382b8627a13c0181c490c96a9b55a5d5c8810c
-assets/packages/cupertino_icons/assets/CupertinoIcons.ttf,1755149637902,12bc6bf55aad4657f62747b6c1c9b5c120a594ed3540db21729b6db3c847340d
-assets/shaders/ink_sparkle.frag,1755149636610,80c6e65c75f1de434b1b22dba61e96ad82dba0f2fc5e8b3b59c2def46d794354
-assets/fonts/MaterialIcons-Regular.otf,1755149637905,b60cf390a06b70ad278b510edc4f8bc3148335df2388dd1993399ff962440173
-assets/NOTICES,1755149636472,42c9f37c8ab682efeefb98072c81dfc334de753e0411d23ffd1d15f692aa4250
-main.dart.js,1755149635888,93008aee12b136aa9f2f721f4fa67f0809162fbd86f4a6be286bee631beb6a84
+.last_build_id,1755493793407,f2403be18940145a0c531d9bb83166b4f10a0d6b482f9713407cd3af781777fe
+flutter_service_worker.js,1755493793103,648d5fadad05120b23e5646676dcd690648328cadf8ae638463489beffb28617
+flutter_bootstrap.js,1755493758680,9bcf25a79a2658757cef895f7e5676474eb829304c9699d67a293a9e87a1bc91
+index.html,1755493758687,691b12b2c1ba770126a5287c22ddb9d30a4d55085a2031eafd1ed93b17bbc64c
+version.json,1755493789691,263c49aaa0e5829f5b7107439b92217c03df18bb3b222c9bee89db582fc5ac00
+assets/AssetManifest.bin,1755493789892,5c90993549e9b9ecd32018eb3ff53f87c4a8c09e4b22cbe11adf0aef287793d8
+assets/AssetManifest.json,1755493789892,b0574e7a18d0015532c01ef15caaf0601fffd6a7cb4fc1348c9e6aa4dc5534ca
+assets/FontManifest.json,1755493789892,9ea504185602e57d97b7c3517d382b8627a13c0181c490c96a9b55a5d5c8810c
+assets/AssetManifest.bin.json,1755493789892,e1fa52e9eef2bd6013745b4e954eeb2cf9a304665f5b4a9fffcf2a5a09f847b3
+assets/packages/cupertino_icons/assets/CupertinoIcons.ttf,1755493792479,12bc6bf55aad4657f62747b6c1c9b5c120a594ed3540db21729b6db3c847340d
+assets/shaders/ink_sparkle.frag,1755493790034,80c6e65c75f1de434b1b22dba61e96ad82dba0f2fc5e8b3b59c2def46d794354
+assets/fonts/MaterialIcons-Regular.otf,1755493792479,a0b6e2f2a56448cf94ce25670c9e66117a38ee020048c54ee7564621b9a3b873
+assets/NOTICES,1755493789892,61732a6f8896626bf1fd392ac46a63fdad371896a6adf8d5d027b3173ed3ab99
+main.dart.js,1755493788623,fb025387e03f6b70cacd540a50a0b6ccac742a38ac2032fe70757624d3a7bee7
